BackendDjango/requirements.txt
asgiref==3.8.1
Django==5.2.3
django-cors-headers==4.7.0
djangorestframework==3.16.0
psycopg2==2.9.10
sqlparse==0.5.3
tzdata==2025.2
django-extensions==3.2.1
Django>=3.2
djangorestframework
# Thêm dòng này
django-cors-headers